backend_bus_app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
geoalchemy2
gpxpy
python-dotenv